Fix normalization to take care of standard elements containing zeroes
Fix normalization to take care of standard elements containing zeroes so that NaN is assigned in the normalized element at the corresponding positions.

This is because the division of the element to normalize between the standard will lead to a division by 0, which is NaN or infinite, in these positions. We have decided to assign both to NaN, indicating that the normalization can't be done in these cases.